feat: create `args` module
